Added openstack and vsphere drivers to node_template resource. go files
