+++ rancher2/resource_rancher2_node_template.go
@@ -147,6 +147,10 @@ func resourceRancher2NodeTemplateUpdate(d *schema.ResourceData, meta interface{}
 		update["azureConfig"] = expandAzureConfig(d.Get("azure_config").([]interface{}))
 	case digitaloceanConfigDriver:
 		update["digitaloceanConfig"] = expandDigitaloceanConfig(d.Get("digitalocean_config").([]interface{}))
+	case openstackConfigDriver:
+		update["openstackConfig"] = expandOpenstackConfig(d.Get("openstack_config").([]interface{}))
+	case vmwarevsphereConfigDriver:
+		update["vmwarevsphereConfig"] = expandVsphereConfig(d.Get("vsphere_config").([]interface{}))
 	}
 
 	newNodeTemplate := &NodeTemplate{}
